refactor(services): introduce ImportExecutor and AudioEffectsController collaborators
Final two follow-ups from PR3 — both files now fall below the 800-LOC
guardrail by extracting genuine collaborator classes rather than just
moving helper functions.

services/StorageAccessService.kt: 1186 -> 645 LOC.
- New ImportExecutor.kt (620 LOC) owns the two large orchestrators
  `executeImportPlanAdvancedImpl` and `importRecursivelyImpl`, plus the
  `getOrCreateLibraryForTypeImpl` helper. They take their DAO / sanitizer
  dependencies as explicit parameters so they can live as top-level
  internal suspend fns in the same package. The Service's public
  methods become thin wrappers that pass `this.libraryDao`,
  `this.mediaItemDao`, `this.metadataDao`, `this.fileNameSanitizer`.

services/music/AdvancedMusicPlayerService.kt: 957 -> 778 LOC.
- New AudioEffectsController.kt (161 LOC, @Singleton + @Inject) owns
  EQ preset, bass-boost, reverb, replay-gain state and persists them
  to its own SharedPreferences. The Service holds it via Hilt and the
  override fns (setEqualizerPreset, enableReverb, setBassBoost,
  setReplayGainEnabled, setReplayGainPreamp, getAudioEffectsSnapshot)
  become single-line delegates.
- New LastFmScrobbleManager.kt (76 LOC, @Singleton + @Inject) owns the
  scrobble coroutine scope, in-flight job, and now-playing/scrobble
  timing. The Service's per-track `startScrobbling` body collapses
  into `scrobbleManager.startScrobbling(track)`.

After this change the only file above the 800-LOC guardrail in this
branch's diff against PR3 is PhosphorIcons.kt, which is addressed by
PR1 (#715) and will drop to 71 LOC once that merges.
